LayoutTests/media/controls/inline-elements-dropoff-order.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../media-file.js"></script>
        <script src="controls-test-helpers.js"></script>
        <script>
            var tester = new ControlsTest("../content/audio-describes-video")
                .whenReady(setup)
                .start();
            var supportsWirelessTargets = !!internals.setMockMediaPlaybackTargetPickerEnabled;
            var runNextStep, currentState, elementNames, testWidthsAndExpectedToBeDropped, testsFinished;

            function setup()
            {
                currentState = tester.currentState;
                tester.test("We are using the apple idiom")
                    .value(currentState.idiom)
                    .isEqualTo("apple");

                if (supportsWirelessTargets)
                    internals.setMockMediaPlaybackTargetPickerState("Wireless playback target", "DeviceNotAvailable");

                runNextStep = runNoAirPlayTests;

                tester.media.addEventListener('webkitplaybacktargetavailabilitychanged', function() {
                    setTimeout(runNextStep, 0);
                });

                elementNames = ["Play Button", "Rewind Button", "Timeline Box", "Mute Box", "AppleTV Device Picker", "Captions Button", "Fullscreen Button"];

                if (!supportsWirelessTargets)
                    setTimeout(runNextStep, 0);
            }

            function runNoAirPlayTests() {
                tester.logBlankLine();
                tester.startNewSection("Without wireless target availability");
                testWidthsAndExpectedToBeDropped = [
                    {
                        width: 310,
                        expectedToBeDropped: [],
                        expectedToBeHidden: ["AppleTV Device Picker"]
                    },
                    {
                        width: 160,
                        expectedToBeDropped: ["Timeline Box"],
                        expectedToBeHidden: ["AppleTV Device Picker"]
                    },
                    {
                        width: 130,
                        expectedToBeDropped: ["Timeline Box", "Captions Button"],
                        expectedToBeHidden: ["AppleTV Device Picker"]
                    },
                    {
                        width: 100,
                        expectedToBeDropped: ["Timeline Box", "Captions Button", "Mute Box"],
                        expectedToBeHidden: ["AppleTV Device Picker"]
                    },
                    {
                        width: 70,
                        expectedToBeDropped: ["Timeline Box", "Captions Button", "Mute Box", "Rewind Button"],
                        expectedToBeHidden: ["AppleTV Device Picker"]
                    },
                    {
                        width: 30,
                        expectedToBeDropped: ["Timeline Box", "Captions Button", "Mute Box", "Rewind Button", "Fullscreen Button"],
                        expectedToBeHidden: ["AppleTV Device Picker"]
                    },
                ];
                testsFinished = function() {
                    if (supportsWirelessTargets)
                        runAirPlayTests();
                    else
                        tester.end();
                };
                setTimeout(runNextTest, 0);
            }

            function runAirPlayTests() {
                tester.logBlankLine();
                tester.startNewSection("With wireless target availability");

                testWidthsAndExpectedToBeDropped = [
                    {
                        width: 350,
                        expectedToBeDropped: [],
                        expectedToBeHidden: []
                    },
                    {
                        width: 200,
                        expectedToBeDropped: ["Timeline Box"],
                        expectedToBeHidden: []
                    },
                    {
                        width: 160,
                        expectedToBeDropped: ["Timeline Box", "AppleTV Device Picker"],
                        expectedToBeHidden: []
                    },
                    {
                        width: 130,
                        expectedToBeDropped: ["Timeline Box", "AppleTV Device Picker", "Captions Button"],
                        expectedToBeHidden: []
                    },
                    {
                        width: 100,
                        expectedToBeDropped: ["Timeline Box", "AppleTV Device Picker", "Captions Button", "Mute Box"],
                        expectedToBeHidden: []
                    },
                    {
                        width: 70,
                        expectedToBeDropped: ["Timeline Box", "AppleTV Device Picker", "Captions Button", "Mute Box", "Rewind Button"],
                        expectedToBeHidden: []
                    },
                    {
                        width: 30,
                        expectedToBeDropped: ["Timeline Box", "AppleTV Device Picker", "Captions Button", "Mute Box", "Rewind Button", "Fullscreen Button"],
                        expectedToBeHidden: []
                    },
                ];

                internals.setMockMediaPlaybackTargetPickerState("Wireless playback target", "DeviceAvailable");
                // Start running the tests after we are told availability has changed.
                runNextStep = runNextTest;

                testsFinished = function() {
                    tester.end();
                };
            }

            function runNextTest() {
                if (!testWidthsAndExpectedToBeDropped.length) {
                    testsFinished();
                    return;
                }

                nextTest = testWidthsAndExpectedToBeDropped.shift();
                tester.media.style.width = nextTest.width + "px";
                setTimeout(testDroppedInlineElements, 0);
            }

            function testDroppedInlineElements() {
                tester.startNewSection(`Testing video at width ${tester.media.style.width}`);
                currentState = tester.currentState;
                elementNames.forEach(function (name) {
                    var elementState = tester.stateForControlsElement(name);

                    function isElementInArray(elementName) {
                        return elementName == name;
                    }

                    var isDropped = elementState.className.indexOf("dropped") > -1;
                    var isHidden = elementState.className.indexOf("hidden") > -1;
                    var expectedToBeDropped = nextTest.expectedToBeDropped.find(isElementInArray);
                    var expectedToBeHidden = nextTest.expectedToBeHidden.find(isElementInArray);
                    if (expectedToBeHidden) {
                        tester.test(`${name} is hidden`)
                            .value(isHidden)
                            .isTrue();
                    } else if (expectedToBeDropped) {
                        tester.test(`${name} is dropped`)
                            .value(isDropped)
                            .isTrue();
                    } else {
                        tester.test(`${name} is visible`)
                            .value(isDropped)
                            .isFalse();
                    }
                });

                setTimeout(runNextTest, 0);
            }

        </script>
        <style>
            body { margin-left: 0; }
            video { width: 310px; height: 80px; }
        </style>
    </head>
    <body>
        <p>This tests that elements in the inline controls drop off when width is restricted, and are in the right place. It also tests that the wireless picker is hidden when no routes are available.</p>
        <p>This test only runs in run-webkit-tests!</p>
        <video controls></video>
    </body>
</html>
