LayoutTests/platform/mac-yosemite/media/controls/inline-elements-dropoff-order-expected.txt

This tests that elements in the inline controls drop off when width is restricted, and are in the right place. It also tests that the wireless picker is hidden when no routes are available.

This test only runs in run-webkit-tests!


EVENT: canplaythrough
PASS: We are using the apple idiom


Without wireless target availability


Testing video at width 310px

PASS: Play Button is visible
PASS: Rewind Button is visible
PASS: Timeline Box is visible
PASS: Mute Box is visible
PASS: AppleTV Device Picker is hidden
PASS: Captions Button is visible
PASS: Fullscreen Button is visible

Testing video at width 160px

PASS: Play Button is visible
PASS: Rewind Button is visible
PASS: Timeline Box is dropped
PASS: Mute Box is visible
PASS: AppleTV Device Picker is hidden
PASS: Captions Button is visible
PASS: Fullscreen Button is visible

Testing video at width 130px

PASS: Play Button is visible
PASS: Rewind Button is visible
PASS: Timeline Box is dropped
PASS: Mute Box is visible
PASS: AppleTV Device Picker is hidden
PASS: Captions Button is dropped
PASS: Fullscreen Button is visible

Testing video at width 100px

PASS: Play Button is visible
PASS: Rewind Button is visible
PASS: Timeline Box is dropped
PASS: Mute Box is dropped
PASS: AppleTV Device Picker is hidden
PASS: Captions Button is dropped
PASS: Fullscreen Button is visible

Testing video at width 70px

PASS: Play Button is visible
PASS: Rewind Button is dropped
PASS: Timeline Box is dropped
PASS: Mute Box is dropped
PASS: AppleTV Device Picker is hidden
PASS: Captions Button is dropped
PASS: Fullscreen Button is visible

Testing video at width 30px

PASS: Play Button is visible
PASS: Rewind Button is dropped
PASS: Timeline Box is dropped
PASS: Mute Box is dropped
PASS: AppleTV Device Picker is hidden
PASS: Captions Button is dropped
PASS: Fullscreen Button is dropped

Testing finished.
